move nested attr helper to utils
