+++ src/anomalib/utils/__init__.py
@@ -2,3 +2,7 @@
 
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
+
+from .attrs import get_nested_attr
+
+__all__ = ["get_nested_attr"]
